fix: handle pr
