+++ storybook/stories/data/index.ts
@@ -6,4 +6,3 @@ export * from './Size';
 export * from './Tick';
 export * from './Time';
 export * from './BoxPlot';
-export * from './Number';
